refactor: packaging (#23)
