+++ src/main/java/dev/saintho/mytly/config/JpaConfig.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 
-@EnableJpaRepositories(basePackages = {"dev.saintho.mytly.repository"})
+@EnableJpaRepositories(basePackages = {"dev.saintho.mytly.repository.jpa"})
 @EnableJpaAuditing
 @Configuration
 public class JpaConfig {
