+++ src/main/java/dev/saintho/mytly/repository/jpa/DailyEngagementRepository.java
@@ -1,4 +1,4 @@
-package dev.saintho.mytly.repository;
+package dev.saintho.mytly.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
